requirements.txt
aiohttp==3.6.0
aiohttp-jinja2==1.1.2
aiohttp-session==2.8.0
emoji==0.5.1
humanize==0.5.1
psycopg2==2.8.3
pymobiledetect==1.3.2
python-dotenv==0.10.1
sqlparse==0.3.0
